refactor: Deleted all unused files and @Deprecated marked ones, bitter updates done on code where it was needed and move some infrastructure contracts out of application layer.
