--- src/main/java/com/jame/dev/gymApp/application/contract/EmailIdentifiable.java
@@ -1,9 +1,0 @@
-package com.jame.dev.gymApp.application.contract;
-
-import com.jame.dev.gymApp.infrastructure.annotation.EmailValid;
-
-import java.util.Optional;
-
-public interface EmailIdentifiable<T> {
-   Optional<T> getByEmail(@EmailValid String email);
-}
