--- src/main/java/com/jame/dev/gymApp/application/contract/EntityRecover.java
@@ -1,5 +1,0 @@
-package com.jame.dev.gymApp.application.contract;
-
-public interface EntityRecover<Resource> {
-   void recover(Resource resource);
-}
