--- src/main/java/com/jame/dev/gymApp/application/contract/EntityRemover.java
@@ -1,5 +1,0 @@
-package com.jame.dev.gymApp.application.contract;
-
-public interface EntityRemover {
-   void hardDelete(long id);
-}
