--- src/main/java/com/jame/dev/gymApp/application/contract/FullService.java
@@ -1,7 +1,0 @@
-package com.jame.dev.gymApp.application.contract;
-
-public interface FullService<DTO_OUT, DTO_IN> extends
-   BaseService<DTO_OUT, DTO_IN>,
-   Patchable<DTO_OUT>,
-   Putable<DTO_OUT, DTO_IN> {
-}
