--- src/main/java/com/jame/dev/gymApp/application/contract/Putable.java
@@ -1,9 +1,0 @@
-package com.jame.dev.gymApp.application.contract;
-
-import com.jame.dev.gymApp.infrastructure.annotation.Minimum;
-import com.jame.dev.gymApp.infrastructure.annotation.NotNullObject;
-import jakarta.validation.Valid;
-
-public interface Putable<E, DTO> {
-   E put(@Minimum long id, @NotNullObject @Valid DTO dto);
-}
