--- src/main/java/com/jame/dev/gymApp/application/contract/Renewable.java
@@ -1,5 +1,0 @@
-package com.jame.dev.gymApp.application.contract;
-
-public interface Renewable<Consumer, Resource> {
-   void applyRenew(Consumer c, Resource r);
-}
