+++ src/main/java/com/jame/dev/gymApp/application/service/IdentityExtractorApplicationService.java
@@ -1,6 +1,6 @@
 package com.jame.dev.gymApp.application.service;
 
-import com.jame.dev.gymApp.application.contract.IdentityExtractorService;
+import com.jame.dev.gymApp.infrastructure.security.principal.IdentityExtractorService;
 import com.jame.dev.gymApp.features.auth.domain.exception.AuthenticationNullException;
 import com.jame.dev.gymApp.features.auth.domain.model.CustomOAuth2User;
 import com.jame.dev.gymApp.features.auth.domain.model.UserPrincipal;
