+++ src/main/java/com/jame/dev/gymApp/features/auth/application/service/AccountRecoveryApplicationService.java
@@ -1,6 +1,6 @@
 package com.jame.dev.gymApp.features.auth.application.service;
 
-import com.jame.dev.gymApp.application.contract.TokenDBHasherService;
+import com.jame.dev.gymApp.infrastructure.security.hash.TokenDBHasherService;
 import com.jame.dev.gymApp.features.auth.application.contract.recovery.AccountRecoveryService;
 import com.jame.dev.gymApp.features.auth.domain.exception.VerificationAttemptFailedException;
 import com.jame.dev.gymApp.features.auth.domain.exception.VerificationNotFoundException;
