+++ src/main/java/com/jame/dev/gymApp/features/auth/application/service/UserDetailsApplicationService.java
@@ -2,10 +2,10 @@ package com.jame.dev.gymApp.features.auth.application.service;
 
 import com.jame.dev.gymApp.domain.exception.NoActiveException;
 import com.jame.dev.gymApp.features.auth.domain.model.UserPrincipal;
-import com.jame.dev.gymApp.features.user.application.contract.UserService;
 import com.jame.dev.gymApp.features.user.application.support.mapper.RoleMapper;
 import com.jame.dev.gymApp.features.user.domain.exception.UserEntityNotFoundException;
 import com.jame.dev.gymApp.features.user.domain.model.UserEntity;
+import com.jame.dev.gymApp.features.user.domain.repository.UserQueryRepository;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -16,13 +16,13 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class UserDetailsApplicationService implements UserDetailsService {
-   private final UserService userService;
+   private final UserQueryRepository userQueryRepository;
    private final RoleMapper roleMapper;
 
    @Override
    @NonNull
    public UserDetails loadUserByUsername(@NonNull String username) throws UsernameNotFoundException {
-      final UserEntity userEntity = userService.getUserByEmail(username)
+      final UserEntity userEntity = userQueryRepository.findByEmail(username)
          .orElseThrow(() -> new UserEntityNotFoundException("User Not Found."));
 
       if (!userEntity.isActive()) {
