+++ src/main/java/com/jame/dev/gymApp/features/auth/application/service/VerificationApplicationService.java
@@ -1,6 +1,6 @@
 package com.jame.dev.gymApp.features.auth.application.service;
 
-import com.jame.dev.gymApp.application.contract.TokenDBHasherService;
+import com.jame.dev.gymApp.infrastructure.security.hash.TokenDBHasherService;
 import com.jame.dev.gymApp.features.auth.application.contract.verification.VerificationService;
 import com.jame.dev.gymApp.features.auth.application.support.factory.VerificationFactory;
 import com.jame.dev.gymApp.features.auth.application.support.helper.VerificationEvaluatorHelper;
