+++ src/main/java/com/jame/dev/gymApp/features/auth/infrastructure/listeners/TokenGenerationSenderListener.java
@@ -1,8 +1,8 @@
 package com.jame.dev.gymApp.features.auth.infrastructure.listeners;
 
 
-import com.jame.dev.gymApp.application.contract.TokenDBHasherService;
-import com.jame.dev.gymApp.application.contract.TokenGeneratorService;
+import com.jame.dev.gymApp.infrastructure.security.hash.TokenDBHasherService;
+import com.jame.dev.gymApp.infrastructure.security.token.TokenGeneratorService;
 import com.jame.dev.gymApp.features.auth.domain.event.TokenGenerationEvent;
 import com.jame.dev.gymApp.features.auth.domain.exception.VerificationNotFoundException;
 import com.jame.dev.gymApp.features.auth.domain.model.VerificationEntity;
