+++ src/main/java/com/jame/dev/gymApp/features/customer/application/contract/CustomerOwnershipService.java
@@ -1,7 +1,7 @@
 package com.jame.dev.gymApp.features.customer.application.contract;
 
 import com.jame.dev.gymApp.features.customer.api.request.CustomerRequest;
-import com.jame.dev.gymApp.application.contract.OwnershipService;
+import com.jame.dev.gymApp.infrastructure.security.owner.OwnershipService;
 import org.jspecify.annotations.NonNull;
 import org.springframework.security.core.Authentication;
