--- src/main/java/com/jame/dev/gymApp/features/customer/application/contract/CustomerRecoverService.java
@@ -1,7 +1,0 @@
-package com.jame.dev.gymApp.features.customer.application.contract;
-
-import com.jame.dev.gymApp.features.auth.api.request.RecoveryRequest;
-import com.jame.dev.gymApp.application.contract.EntityRecover;
-
-public interface CustomerRecoverService extends EntityRecover<RecoveryRequest> {
-}
