--- src/main/java/com/jame/dev/gymApp/features/user/application/contract/UserInactiveService.java
@@ -1,7 +1,0 @@
-package com.jame.dev.gymApp.features.user.application.contract;
-
-import com.jame.dev.gymApp.features.user.api.response.UserMinimalInfoResponse;
-import com.jame.dev.gymApp.application.contract.InactiveEntityService;
-
-public interface UserInactiveService extends InactiveEntityService<UserMinimalInfoResponse> {
-}
