+++ src/main/java/com/jame/dev/gymApp/infrastructure/security/principal/IdentityExtractorService.java
@@ -1,4 +1,4 @@
-package com.jame.dev.gymApp.application.contract;
+package com.jame.dev.gymApp.infrastructure.security.principal;
 
 import com.jame.dev.gymApp.infrastructure.annotation.NotNullObject;
 import org.springframework.security.core.Authentication;
