--- src/main/java/com/jame/dev/gymApp/infrastructure/web/BaseController.java
@@ -1,68 +1,0 @@
-package com.jame.dev.gymApp.infrastructure.web;
-
-import com.jame.dev.gymApp.infrastructure.annotation.Minimum;
-import com.jame.dev.gymApp.infrastructure.annotation.NotNullObject;
-import com.jame.dev.gymApp.application.dto.PageDto;
-import com.jame.dev.gymApp.application.contract.BaseService;
-import jakarta.validation.Valid;
-import lombok.NonNull;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.Pageable;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
-import java.net.URI;
-import java.util.function.Function;
-
-@Validated
-public abstract class BaseController<OUT, IN> {
-
-   protected final BaseService<OUT, IN> service;
-   protected final Function<OUT, Long> idExtractor;
-
-   protected BaseController(
-      BaseService<OUT, IN> service,
-      Function<OUT, Long> idExtractor) {
-      this.service = service;
-      this.idExtractor = idExtractor;
-   }
-
-   protected ResponseEntity<@NonNull Page<@NonNull OUT>> getPage(
-      final Pageable pageable, final String search) {
-      final PageDto<@NonNull OUT> dtoPage = service.getPage(pageable, search);
-      final Page<OUT> response = new PageImpl<>(dtoPage.content(), pageable, dtoPage.totalElements());
-      return ResponseEntity.ok(response);
-   }
-
-   protected ResponseEntity<@NonNull OUT> getOne(final long id) {
-      final OUT body = service.getById(id);
-      return ResponseEntity.ok(body);
-   }
-
-   protected ResponseEntity<@NonNull OUT> create(final IN dto) {
-      final OUT response = service.save(dto);
-      final URI created = ServletUriComponentsBuilder
-         .fromCurrentRequest()
-         .path("/{id}")
-         .buildAndExpand(idExtractor.apply(response))
-         .toUri();
-      return ResponseEntity.created(created)
-         .body(response);
-   }
-
-   protected ResponseEntity<@NonNull OUT> update(
-      @Minimum final long id,
-      @Valid
-      @NotNullObject(message = "Payload must not be null") final IN dto) {
-      final OUT response = service.update(id, dto);
-      return ResponseEntity.ok(response);
-   }
-
-   protected ResponseEntity<Void> delete(
-      @Minimum final long id) {
-      service.softDelete(id);
-      return ResponseEntity.noContent().build();
-   }
-}
