+++ src/test/java/com/jame/dev/gymApp/service/VerificationServiceTest.java
@@ -1,6 +1,6 @@
 package com.jame.dev.gymApp.service;
 
-import com.jame.dev.gymApp.application.contract.TokenDBHasherService;
+import com.jame.dev.gymApp.infrastructure.security.hash.TokenDBHasherService;
 import com.jame.dev.gymApp.features.auth.application.service.VerificationApplicationService;
 import com.jame.dev.gymApp.features.auth.application.support.factory.VerificationFactory;
 import com.jame.dev.gymApp.features.auth.application.support.helper.VerificationEvaluatorHelper;
